Fix AFESamnaPDM implementation
Add a readregister call in between the flipped and real data.
Then we know all the events that can be dropped.
Start the stopwatch only when sending the real data.
Fix a bug with the dt size (when using flip and encode, dt was not taking
into consideration the doubling of the input) and the flipped data was being
added to the array but not written into the device (no buffer.write() call).