Remove __del__ from XyloMonitor
In the destructor, we weres switching off power measurement,
which is not an issue per se, however, without an explicit call
to the destructor, this was causing weird behaviors with the garbage
collector, making it being called at odd times.

Now, the activation of power measurement verify if it was active
and then deactivates and activates is again (it needs to be done cause
if a new instance of the class is created, the previous activation doesnt
work).

Also, update tutorials to not call `del` explictly anymore.